requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.62.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.13
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
